requirements.txt
django-redactoreditor
